refactor: Move timeout helpers to new internal lib
